bsp fixed

+++ C02/ex03/src/bsp.cpp
@@ -2,21 +2,21 @@
 #include "Fixed.hpp"
 #include <cstdio>
 
-bool haveSameSign(int x, int y);
-int	cross_product_2d(int x[], int y[], int z[]);
+bool haveSameSign(float x, float y);
+float cross_product_2d(float x[], float y[], float z[]);
 void	initializeArray(int array[], Point const p);
 
-int	cross_product_2d(int x[], int y[], int z[]){
+float	cross_product_2d(float x[], float y[], float z[]){
 	return ((y[0]-x[0])*(z[1]-x[1])-(y[1]-x[1])*(z[0]-x[0]));
 }
 
-void	initializeArray(int array[], Point const p)
+void	initializeArray(float array[], Point const p)
 {
-	array[0] = p.get_x().toInt();
-	array[1] = p.get_y().toInt();
+	array[0] = p.get_x().toFloat();
+	array[1] = p.get_y().toFloat();
 }
 
-bool haveSameSign(int x, int y)
+bool haveSameSign(float x, float y)
 {
 	if ((x > 0 && y > 0) || (x < 0 && y < 0))
 		return (true);
@@ -30,17 +30,17 @@ bool haveSameSign(int x, int y)
  *========================================================================**/
 bool bsp( Point const a, Point const b, Point const c, Point const point){
 	
-	int A[2], B[2], C[2], P[2];
+	float A[2], B[2], C[2], P[2];
 	
 	initializeArray(A, a);
 	initializeArray(B, b);
 	initializeArray(C, c);
 	initializeArray(P, point);
 
-	int cp1 = cross_product_2d(A, B, C);
-	int cp2 = cross_product_2d(A, B, P);
-	int cp3 = cross_product_2d(B, C, P);
-	int cp4 = cross_product_2d(C, A, P);
+	float cp1 = cross_product_2d(A, B, C);
+	float cp2 = cross_product_2d(A, B, P);
+	float cp3 = cross_product_2d(B, C, P);
+	float cp4 = cross_product_2d(C, A, P);
 
 	bool sameSign1 = haveSameSign(cp1, cp2);
 	bool sameSign2 = haveSameSign(cp1, cp3);
